+++ clang/docs/ReleaseNotes.rst
@@ -160,6 +160,12 @@ X86 Support in Clang
   ``_popcnt64``, ``__popcntd`` and ``__popcntq``  may now be used within
   constexpr expressions.
 
+- The x86 intrinsics ``_bit_scan_forward``, ``__bsfd`` and ``__bsfq`` may now
+  be used within constexpr expressions.
+
+- The x86 intrinsics ``_bit_scan_reverse``, ``__bsrd`` and ``__bsrq`` may now
+  be used within constexpr expressions.
+
 - The x86 intrinsics ``__bswap``, ``__bswapd``, ``__bswap64`` and ``__bswapq``
   may now be used within constexpr expressions.